feat: generateMetadata
